+++ CPAC/pipeline/engine.py
@@ -1625,6 +1625,55 @@ def ingress_raw_anat_data(wf, rpool, cfg, data_paths, unique_id, part_id,
         rpool.set_data('T2w', anat_flow_T2, 'outputspec.anat', {},
                     "", "anat_ingress")
 
+    if 'freesurfer_dir' in data_paths['anat']:
+        anat['freesurfer_dir'] = data_paths['anat']['freesurfer_dir']
+
+        fs_ingress = create_general_datasource('gather_freesurfer_dir')
+        fs_ingress.inputs.inputnode.set(
+            unique_id=unique_id,
+            data=data_paths['anat']['freesurfer_dir'],
+            creds_path=data_paths['creds_path'],
+            dl_dir=cfg.pipeline_setup['working_directory']['path'])
+        rpool.set_data("freesurfer-subject-dir", fs_ingress, 'outputspec.data',
+                       {}, "", "freesurfer_config_ingress")
+
+        recon_outs = {
+            'raw-average': 'mri/rawavg.mgz',
+            'subcortical-seg': 'mri/aseg.mgz',
+            'brainmask': 'mri/brainmask.mgz',
+            'wmparc': 'mri/wmparc.mgz',
+            'T1': 'mri/T1.mgz',
+            'hemi-L_desc-surface_curv': 'surf/lh.curv',
+            'hemi-R_desc-surface_curv': 'surf/rh.curv',
+            'hemi-L_desc-surfaceMesh_pial': 'surf/lh.pial',
+            'hemi-R_desc-surfaceMesh_pial': 'surf/rh.pial',
+            'hemi-L_desc-surfaceMesh_smoothwm': 'surf/lh.smoothwm',
+            'hemi-R_desc-surfaceMesh_smoothwm': 'surf/rh.smoothwm',
+            'hemi-L_desc-surfaceMesh_sphere': 'surf/lh.sphere',
+            'hemi-R_desc-surfaceMesh_sphere': 'surf/rh.sphere',
+            'hemi-L_desc-surfaceMap_sulc': 'surf/lh.sulc',
+            'hemi-R_desc-surfaceMap_sulc': 'surf/rh.sulc',
+            'hemi-L_desc-surfaceMap_thickness': 'surf/lh.thickness',
+            'hemi-R_desc-surfaceMap_thickness': 'surf/rh.thickness',
+            'hemi-L_desc-surfaceMap_volume': 'surf/lh.volume',
+            'hemi-R_desc-surfaceMap_volume': 'surf/rh.volume',
+            'hemi-L_desc-surfaceMesh_white': 'surf/lh.white',
+            'hemi-R_desc-surfaceMesh_white': 'surf/rh.white',
+        }
+        
+        for key, outfile in recon_outs.items():
+            fullpath = os.path.join(data_paths['anat']['freesurfer_dir'],
+                                    outfile)
+            if os.path.exists(fullpath):
+                fs_ingress = create_general_datasource(f'gather_fs_{key}_dir')
+                fs_ingress.inputs.inputnode.set(
+                    unique_id=unique_id,
+                    data=fullpath,
+                    creds_path=data_paths['creds_path'],
+                    dl_dir=cfg.pipeline_setup['working_directory']['path'])
+                rpool.set_data(key, fs_ingress, 'outputspec.data',
+                               {}, "", f"fs_{key}_ingress")
+
     return rpool
 
 
@@ -1942,17 +1991,6 @@ def ingress_pipeconfig_paths(cfg, rpool, unique_id, creds_path=None):
                 )
                 rpool.set_data(key, config_ingress, 'outputspec.data',
                                json_info, "", f"{key}_config_ingress")
-
-    # Freesurfer directory, not a template, so not in cpac_templates.tsv
-    if cfg.surface_analysis['freesurfer']['freesurfer_dir']:
-        fs_ingress = create_general_datasource('gather_freesurfer_dir')
-        fs_ingress.inputs.inputnode.set(
-            unique_id=unique_id,
-            data=cfg.surface_analysis['freesurfer']['freesurfer_dir'],
-            creds_path=creds_path,
-            dl_dir=cfg.pipeline_setup['working_directory']['path'])
-        rpool.set_data("freesurfer-subject-dir", fs_ingress, 'outputspec.data',
-                       json_info, "", "freesurfer_config_ingress")
 
     # templates, resampling from config
     '''