+++ CPAC/pipeline/schema.py
@@ -710,9 +710,8 @@ latest_schema = Schema({
     },
     'surface_analysis': {
         'freesurfer': {
-            'run': bool1_1,
-            'reconall_args': Maybe(str),
-            'freesurfer_dir': Maybe(str)
+            'run_reconall': bool1_1,
+            'reconall_args': Maybe(str)
         },
         'post_freesurfer': {
             'run': bool1_1,